requirements.txt
flask
pyyaml
joblib
scikit-learn
spacy